# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: UGNAY Research Continuity API
  version: 0.2.0
  description: Explainable research discovery, traceability, scope analysis, and continuity API. Recommendations never replace academic decisions.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /auth/login:
    post:
      security: []
      summary: Start a same-origin server session
      description: Used by the browser account panel after `/auth/me` reports a live anonymous session. A successful response establishes the HttpOnly JDBC-backed session cookie; no browser access token is returned or stored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { description: Authenticated session, content: { application/json: { schema: { $ref: '#/components/schemas/AuthView' } } } }
        '401': { $ref: '#/components/responses/Problem' }
...
  /auth/logout:
    post:
      security: []
      summary: Invalidate the current JDBC-backed session
      description: The browser account panel first obtains `/auth/csrf`, sends the named token header on this request, then invalidates and refetches session and workspace queries.
      parameters: [ { $ref: '#/components/parameters/CsrfHeader' } ]
      responses:
        '204': { description: Session invalidated }
        '403': { $ref: '#/components/responses/Problem' }
  /auth/invitations/accept:
    post:
      security: []
      summary: Accept a single-use invitation and create a persistent account
      description: The invitation token authorizes this public action and must be unused and unexpired. The account inherits the inviter's department and receives the invitation's one intended role; acceptance does not start a login session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, displayName, password]
              properties:
                token: { type: string, minLength: 1 }
                displayName: { type: string, minLength: 1, maxLength: 160 }
                password: { type: string, format: password, minLength: 12, maxLength: 128 }
      responses:
        '201': { description: Persistent invited account, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /imports/documents:
    post:
      summary: Validate, scan, privately store, and durably queue PDF extraction
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties: { file: { type: string, format: binary } }
      responses:
        '202': { description: "Durable extraction job ID, status URL, and progress-events URL" }
        '422': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /imports/documents/jobs/{jobId}:
    get:
      summary: Read persistent PDF extraction status, limits, progress, and publication eligibility
      parameters: [ { name: jobId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: Persistent extraction job }, '404': { $ref: '#/components/responses/Problem' } }
  /imports/documents/jobs/{jobId}/events:
    get:
      summary: Subscribe to extraction progress while GET status remains the durable source of truth
      parameters: [ { name: jobId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '200': { description: text/event-stream progress events }, '404': { $ref: '#/components/responses/Problem' } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  parameters:
...
    CsrfHeader: { in: header, name: X-XSRF-TOKEN, required: true, schema: { type: string }, description: Token returned by GET /auth/csrf and paired with its session cookie }
  responses:
    Problem:
      description: RFC 9457 problem detail
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    AuthView:
      type: object
      required: [authenticated, roles]
      properties:
        authenticated: { type: boolean }
        email: { type: [string, 'null'] }
        roles: { type: array, items: { type: string } }
    User:
      type: object
      required: [id, email, displayName, status, roles]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        department: { type: [string, 'null'] }
        status: { type: string }
        roles: { type: array, items: { enum: [STUDENT, ADVISER, COORDINATOR, CURATOR] } }
...
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }